Fix argument type mismatch warning
- another integer/NSInteger mismatch.

+++ ControlPanel.m
@@ -31,7 +31,7 @@
 @implementation ControlPanel
 
 - (id)initWithContentRect:(NSRect)contentRect
-                styleMask:(unsigned int)styleMask
+                styleMask:(NSUInteger)styleMask
                   backing:(NSBackingStoreType)bufferingType
                     defer:(BOOL)deferCreation
 {
